pyproject.toml
[tool.poetry]
name = "construction-site"
version = "0.1.0"
description = ""
authors = ["macbook <kriukov@pos-credir.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"
mypy = "^1.15.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
v = "^1"
redis = "^5.2.1"
ipython = "^9.0.2"
gunicorn = "^23.0.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
